Textbook-like plot for squared error cost function.
I've fixed my errors (not using meshgrids for a surface plot)
and used my own code - without copypasting from the Labs - to do the plot.
Now it looks like the one from the textbook. Just remember that this plot
shows that Squared Error Cost Function is bad for the classification.